+++ src/app/product/[id]/Detail.tsx
@@ -1,8 +1,26 @@
 'use client';
 
 import Loading from '@/app/loading';
+import type {
+  InventoryItem,
+  OptionHandlers,
+  ProductItem,
+  SelectedOption,
+} from '@/types/product';
 import { useCartManager, useSelectedOptions } from './hooks';
 import { DetailView } from './ui';
+
+export type DetailViewProps = DetailViewBase & OptionHandlers;
+
+interface DetailViewBase {
+  item: ProductItem;
+  price: number;
+  inventory: InventoryItem[];
+  totalQuantity: number;
+  selectedOptions: SelectedOption[];
+  onSelectSize: (size: number) => () => void;
+  onCartButton: () => void;
+}
 
 export default function Detail() {
   const {
